Databases tested, file stuff improved

+++ database/src/commonMain/kotlin/com/lightningkite/services/database/numplus.kt
@@ -4,8 +4,8 @@ package com.lightningkite.services.database
 @Suppress("UNCHECKED_CAST")
 internal operator fun <T : Number> T.plus(other: T): T {
     return when (this) {
-        is Byte -> (this + other as Byte) as T
-        is Short -> (this + other as Short) as T
+        is Byte -> (this + other as Byte).toByte() as T
+        is Short -> (this + other as Short).toShort() as T
         is Int -> (this + other as Int) as T
         is Long -> (this + other as Long) as T
         is Float -> (this + other as Float) as T
@@ -17,8 +17,8 @@ internal operator fun <T : Number> T.plus(other: T): T {
 @Suppress("UNCHECKED_CAST")
 internal operator fun <T : Number> T.times(other: T): T {
     return when (this) {
-        is Byte -> (this * other as Byte) as T
-        is Short -> (this * other as Short) as T
+        is Byte -> (this * other as Byte).toByte() as T
+        is Short -> (this * other as Short).toShort() as T
         is Int -> (this * other as Int) as T
         is Long -> (this * other as Long) as T
         is Float -> (this * other as Float) as T
